// src/schemas/ts-rest.ts
import { z } from 'zod';

export const TsRestErrorSchema = z.object({
  message: z.string(),
  code: z.string().optional(),
  params: z.record(z.string(), z.unknown()).optional(),
  details: z.unknown().optional(),
});

export type TsRestError = z.infer<typeof TsRestErrorSchema>;

export type TsRestMetaData =
  | { auth: 'public' }
  | {
      auth: 'required';
      permissionKey?: {
        resourceKey: string;
        actionKey: string;
      };
    };